pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "tqdm>=4.67.1",
]
